Move PatternRecognition to plugin
