+++ RecoHGCal/TICL/interface/PatternRecognitionAlgoBase.h
